Extract styles to separate css files
