+++ src/components/Header/styles.css
@@ -1,0 +1,9 @@
+.header {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+
+    box-shadow: 1px 8px 13px -3px rgba(41,41,41,0.51);
+    -webkit-box-shadow: 1px 8px 13px -3px rgba(41,41,41,0.51);
+    -moz-box-shadow: 1px 8px 13px -3px rgba(41,41,41,0.51);
+}
