+++ src/components/Pagination/index.tsx
@@ -1,8 +1,8 @@
 import React, { useState } from "react";
 import { Pagination, PaginationProps } from "antd";
 import { useDispatch, useSelector } from "react-redux";
-import { AppDispatch, RootState } from "../store";
-import { fetchPokemons } from "../store/pokemonSlicer";
+import { AppDispatch, RootState } from "../../store";
+import { fetchPokemons } from "../../store/pokemonSlicer";
 
 const CustomPagination: React.FC = () => {
   const [currentPage, setCurrentPage] = useState(0);
