+++ src/components/PokemonCard/index.tsx
@@ -1,7 +1,8 @@
 import React from "react";
 import { Card, Image, Typography, Flex } from "antd";
 import { Link } from "react-router-dom";
-import emptyImg from "../../assets/no-pictures.png";
+import emptyImg from "../../../assets/no-pictures.png";
+import "./styles.css";
 
 const { Title, Text } = Typography;
 
@@ -10,39 +11,21 @@ export type PokemonBaseInfo = {
     id: number;
     name: string;
   };
-  isDesktop?: boolean;
 };
 
-const PokemonCard: React.FC<PokemonBaseInfo> = ({ character, isDesktop }) => {
+const PokemonCard: React.FC<PokemonBaseInfo> = ({ character }) => {
   const { id, name } = character;
   return (
     <Link to={`/pokemon/${name}`}>
-      <Card
-        style={{
-          width: isDesktop ? 300 : "90vw",
-          minHeight: "100%",
-          textAlign: "center",
-        }}
-        hoverable
-      >
+      <Card className="pokemon-card" hoverable>
         <Image
           preview={false}
           src={`https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/${id}.png`}
           fallback={emptyImg}
         />
         <Flex justify="center" align="center" vertical>
-          <Text type="secondary" style={{ marginTop: "8px" }}>
-            #{id}
-          </Text>
-          <Title
-            level={4}
-            style={{
-              margin: "0",
-              textTransform: "capitalize",
-              fontWeight: 700,
-              color: "#515052",
-            }}
-          >
+          <Text type="secondary">#{id}</Text>
+          <Title level={4} className="pokemon-card-name">
             {name}
           </Title>
         </Flex>
