+++ src/components/PokemonsList/index.tsx
@@ -1,31 +1,22 @@
 import React from "react";
 import { useSelector } from "react-redux";
-import { Spin, Row, Col, Grid, Empty } from "antd";
-import { RootState } from "../store";
-import PokemonCard from "./PokemonCard";
-
-const { useBreakpoint } = Grid;
+import { Spin, Row, Col, Empty } from "antd";
+import { RootState } from "../../store";
+import PokemonCard from "../PokemonCard";
+import "./styles.css";
 
 const PokemonsList: React.FC = () => {
   const pokemonState = useSelector((state: RootState) => state.pokemon);
-  const screens = useBreakpoint();
-  const isDesktop = screens.sm;
 
   return (
     <div>
       {pokemonState.loading && <Spin tip="Loading" size="large" fullscreen />}
 
       {!pokemonState.error ? (
-        <Row
-          style={{
-            marginBottom: "40px",
-          }}
-          gutter={[16, 16]}
-          justify="center"
-        >
+        <Row className="cards-row" gutter={[16, 16]} justify="center">
           {pokemonState.data.map((item: { id: number; name: string }) => (
             <Col key={item.id}>
-              <PokemonCard character={item} isDesktop={isDesktop} />
+              <PokemonCard character={item} />
             </Col>
           ))}
         </Row>
