+++ src/components/PokemonsList/styles.css
@@ -1,0 +1,3 @@
+.cards-row {
+  margin-bottom: 40px;
+}
