+++ src/routes/Character/styles.css
@@ -1,0 +1,30 @@
+.breadcrumb {
+  margin-bottom: 40px;
+  align-self: flex-start;
+}
+
+.pokemon-avatar {
+  background-color: #fff;
+  box-shadow: 4px 8px 13px -4px rgba(41, 41, 41, 0.39);
+  -webkit-box-shadow: 4px 8px 13px -4px rgba(41, 41, 41, 0.39);
+  -moz-box-shadow: 4px 8px 13px -4px rgba(41, 41, 41, 0.39);
+}
+
+.pokemon-name {
+  color: #515052 !important;
+  text-transform: capitalize;
+  margin-top: 24px;
+
+  @media (min-width: 576px) {
+    margin-left: 40px;
+  }
+}
+
+.info-header {
+  width: 100%;
+  flex-direction: column;
+
+  @media (min-width: 576px) {
+    flex-direction: row;
+  }
+}
