+++ src/routes/Home/index.tsx
@@ -1,21 +1,19 @@
 import React, { useEffect } from "react";
-import { Typography, Input, Grid } from "antd";
+import { Typography, Input } from "antd";
 import { useDispatch } from "react-redux";
 import { SearchProps } from "antd/es/input";
-import PokemonsList from "../components/PokemonsList";
-import { fetchPokemons, searchPokemon } from "../store/pokemonSlicer";
-import { AppDispatch } from "../store";
-import CustomPagination from "../components/Pagination";
-import Page from "../components/Page";
+import PokemonsList from "../../components/PokemonsList";
+import { fetchPokemons, searchPokemon } from "../../store/pokemonSlicer";
+import { AppDispatch } from "../../store";
+import CustomPagination from "../../components/Pagination";
+import Page from "../../components/Page";
+import "./styles.css";
 
 const { Title, Text } = Typography;
 const { Search } = Input;
-const { useBreakpoint } = Grid;
 
 const Home: React.FC = () => {
   const dispatch = useDispatch<AppDispatch>();
-  const screens = useBreakpoint();
-  const isDesktop = screens.md;
 
   const handleSearch: SearchProps["onSearch"] = (value) => {
     const formattedValue = value.toLowerCase().trim();
@@ -34,15 +32,12 @@ const Home: React.FC = () => {
   return (
     <Page>
       <Title>Welcome to Pokédex!</Title>
-      <Text style={{ marginBottom: "24px" }}>
+      <Text className="intro">
         Enter a Pokémon's name or number in the search bar below to begin your
         journey.
       </Text>
       <Search
-        style={{
-          width: `${isDesktop ? "50%" : "100%"}`,
-          marginBottom: "40px",
-        }}
+        className="search-bar"
         placeholder="Pokémon's name or number"
         enterButton="Search"
         size="large"
